Account checks and hardening
Add IBM Cloud account checks to the login path. This in turn requires the account and administrator API key always be configured on the auth method.

The administrator token is now used for all IBM Cloud API calls except the call to fetch a user or service ID token using an API key.

The module, code package, auth backend struct name have been changed to be more unique and descriptive.