Refactor http comm methods and add login path tests
Encapsulate the methods IAM helper and token methods that communicate with
IBM Cloud into an interface and implementing struct.

This encapsulation provides a logical grouping of functions, but also allows the
creation of a mock IAM helper which allows for success and error path testing
for the login path logic without the need to configure a test environment in
the cloud. This commit delivers a login path test implementation that provides this.